refactor(rate-of-closure): split launch monitor analytics modules
